refactor: 체험 등록/수정 폼 이미지 영역 컴포넌트로 분리
