+++ src/hooks/useMyActivityForm.tsx
@@ -1,6 +1,6 @@
 import { useFieldArray, useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { MyActivitySchema } from '@/lib/utils/myActivitySchema';
+import { MyActivityFormData, MyActivitySchema } from '@/lib/utils/myActivitySchema';
 import { useMutation } from '@tanstack/react-query';
 import {
   ActivityCreateRequest,
@@ -13,10 +13,19 @@ import {
   updateActivity,
   uploadActivityImage,
 } from '@/app/api/activities';
-import { ActivityDetail } from '@/types/activities.type';
-import { errorToast, successToast } from '@/lib/utils/toastUtils';
+import { ActivitiesCategoryType, ActivityDetail } from '@/types/activities.type';
+import { successToast } from '@/lib/utils/toastUtils';
+import { toApiDate } from '@/lib/utils/dateUtils';
+import { useState } from 'react';
 
-const useMyActivityForm = () => {
+interface useMyActivityForm {
+  mode?: 'EDIT' | 'REGISTER';
+  activityId?: string;
+}
+
+const useMyActivityForm = ({ mode = 'REGISTER', activityId }: useMyActivityForm) => {
+  const [originalSchedules, setOriginalSchedules] = useState<MyActivityFormData['schedules']>([]);
+
   const methods = useForm({
     resolver: zodResolver(MyActivitySchema),
     defaultValues: {
@@ -107,6 +116,101 @@ const useMyActivityForm = () => {
     },
   });
 
+  const uploadImageAndGetUrl = async () => {
+    const bannerFiles = watch('bannerFiles') ?? [];
+    const subFiles = watch('subFiles') ?? [];
+
+    if (bannerFiles.length > 0) {
+      const bannerUploadResponse = await uploadImageMutation.mutateAsync(bannerFiles[0]);
+      setValue('bannerImageUrl', bannerUploadResponse.activityImageUrl, {
+        shouldValidate: true,
+        shouldDirty: true,
+      });
+    }
+
+    if (subFiles.length > 0) {
+      const subImagesUploadResponse = await Promise.all(
+        subFiles.map((file) => uploadImageMutation.mutateAsync(file)),
+      );
+
+      const subImageUrls = subImagesUploadResponse.map((res) => res.activityImageUrl);
+
+      if (mode === 'EDIT') {
+        setValue('subImageUrlsToAdd', subImageUrls, { shouldValidate: true, shouldDirty: true });
+      } else {
+        setValue('subImageUrls', subImageUrls, { shouldValidate: true, shouldDirty: true });
+      }
+    }
+  };
+
+  const registerForm = async () => {
+    const values = methods.getValues();
+
+    const params: ActivityCreateRequest = {
+      ...values,
+      category: values.category as ActivitiesCategoryType,
+      price: Number(values.price.replace(/,/g, '')),
+      subImageUrls: values.subImageUrls ?? [],
+      schedules: values.schedules.map(({ date, startTime, endTime }) => ({
+        date: toApiDate(date),
+        startTime,
+        endTime,
+      })),
+    };
+
+    registerMutation.mutate(params);
+  };
+
+  const updateForm = async () => {
+    const values = methods.getValues();
+
+    // 삭제된 스케줄
+    const scheduleIdsToRemove: number[] = originalSchedules
+      .filter((s) => {
+        const current = values.schedules.find((v) => v.id === s.id);
+        if (!current) return true;
+
+        return (
+          current.date !== s.date ||
+          current.startTime !== s.startTime ||
+          current.endTime !== s.endTime
+        );
+      })
+      .map((s) => Number(s.id));
+
+    // 새로 추가된 스케줄
+    const schedulesToAdd = values.schedules
+      .filter((s) => {
+        if (!s.id) return true;
+
+        const original = originalSchedules.find((v) => v.id === s.id);
+        if (!original) return true;
+
+        return (
+          original.date !== s.date ||
+          original.startTime !== s.startTime ||
+          original.endTime !== s.endTime
+        );
+      })
+      .map(({ date, startTime, endTime }) => ({
+        date: toApiDate(date),
+        startTime,
+        endTime,
+      }));
+
+    const params: ActivityUpdateRequest = {
+      ...values,
+      category: values.category as ActivitiesCategoryType,
+      price: Number(values.price.replace(/,/g, '')),
+      scheduleIdsToRemove,
+      schedulesToAdd,
+      subImageIdsToRemove: values.subImageIdsToRemove ?? [],
+      subImageUrlsToAdd: values.subImageUrlsToAdd ?? [],
+    };
+
+    updateMutation.mutate({ activityId: Number(activityId), data: params });
+  };
+
   return {
     methods,
     errors,
@@ -124,8 +228,10 @@ const useMyActivityForm = () => {
     uploadImageMutation,
     registerMutation,
     updateMutation,
-    successToast,
-    errorToast,
+    uploadImageAndGetUrl,
+    registerForm,
+    updateForm,
+    setOriginalSchedules,
   };
 };
